// src/common/enums/payment.enum.ts
export enum PaymentStatus {
  PENDIENTE = 'PENDIENTE',
  PAGADO = 'PAGADO',
  VENCIDO = 'VENCIDO',
  PARCIAL = 'PARCIAL',
  CANCELADO = 'CANCELADO',
}

export enum PaymentMethod {
  TRANSFERENCIA = 'TRANSFERENCIA',
  EFECTIVO = 'EFECTIVO',
  TARJETA = 'TARJETA',
  DEBITO_AUTOMATICO = 'DEBITO_AUTOMATICO',
}